RegistrationService/requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
passlib==1.7.4
bcrypt==3.2.0
python-jose[cryptography]
python-multipart
pydantic
bson
